+++ storage/src/storage/mod.rs
@@ -39,8 +39,8 @@ pub use local::XvcLocalStorage;
 use serde::{Deserialize, Serialize};
 use tempfile::TempDir;
 use uuid::Uuid;
-use xvc_core::{error, XvcOutputSender};
 use xvc_core::AbsolutePath;
+use xvc_core::{error, XvcOutputSender};
 
 use clap_complete::CompletionCandidate;
 
@@ -48,8 +48,8 @@ use crate::{Error, Result, StorageIdentifier};
 
 use relative_path::{RelativePath, RelativePathBuf};
 
-use xvc_core::{util::completer::load_store_for_completion, XvcCachePath, XvcRoot};
 use xvc_core::{persist, XvcStore};
+use xvc_core::{util::completer::load_store_for_completion, XvcCachePath, XvcRoot};
 
 use self::generic::XvcGenericStorage;
 
@@ -134,6 +134,52 @@ impl XvcStorage {
             XvcStorage::Wasabi(s) => s.guid.to_string(),
             #[cfg(feature = "digital-ocean")]
             XvcStorage::DigitalOcean(s) => s.guid.to_string(),
+        }
+    }
+
+    /// Return a dynamic reference to the underlying storage
+    pub fn as_dyn(&self) -> &dyn XvcStorageOperations {
+        match self {
+            XvcStorage::Local(s) => s,
+            XvcStorage::Generic(s) => s,
+            XvcStorage::Rsync(s) => s,
+            #[cfg(feature = "rclone")]
+            XvcStorage::Rclone(s) => s,
+            #[cfg(feature = "s3")]
+            XvcStorage::S3(s) => s,
+            #[cfg(feature = "minio")]
+            XvcStorage::Minio(s) => s,
+            #[cfg(feature = "r2")]
+            XvcStorage::R2(s) => s,
+            #[cfg(feature = "gcs")]
+            XvcStorage::Gcs(s) => s,
+            #[cfg(feature = "wasabi")]
+            XvcStorage::Wasabi(s) => s,
+            #[cfg(feature = "digital-ocean")]
+            XvcStorage::DigitalOcean(s) => s,
+        }
+    }
+
+    /// Return a mutable dynamic reference to the underlying storage
+    pub fn as_dyn_mut(&mut self) -> &mut dyn XvcStorageOperations {
+        match self {
+            XvcStorage::Local(s) => s,
+            XvcStorage::Generic(s) => s,
+            XvcStorage::Rsync(s) => s,
+            #[cfg(feature = "rclone")]
+            XvcStorage::Rclone(s) => s,
+            #[cfg(feature = "s3")]
+            XvcStorage::S3(s) => s,
+            #[cfg(feature = "minio")]
+            XvcStorage::Minio(s) => s,
+            #[cfg(feature = "r2")]
+            XvcStorage::R2(s) => s,
+            #[cfg(feature = "gcs")]
+            XvcStorage::Gcs(s) => s,
+            #[cfg(feature = "wasabi")]
+            XvcStorage::Wasabi(s) => s,
+            #[cfg(feature = "digital-ocean")]
+            XvcStorage::DigitalOcean(s) => s,
         }
     }
 }
@@ -224,7 +270,7 @@ pub struct XvcStorageTempDir(AbsolutePath);
 impl XvcStorageTempDir {
     /// Create a new temporary directory
     pub fn new() -> Result<Self> {
-        let temp_dir = AbsolutePath::from(TempDir::new()?.into_path());
+        let temp_dir = AbsolutePath::from(TempDir::new()?.keep());
         Ok(Self(temp_dir))
     }